Correction pour le code duplicated
