+++ Application/src/environment/themes.tsx
@@ -1,50 +1,4 @@
-import { MD3LightTheme, MD3DarkTheme, MD3Theme } from 'react-native-paper';
-
-export const lightTheme: MD3Theme = {
-  ...MD3LightTheme,
-  colors: {
-    primary: 'rgb(6, 35, 51)',
-    onPrimary: 'rgb(237, 246, 253)',
-    primaryContainer: 'rgb(200, 230, 255)',
-    onPrimaryContainer: 'rgb(0, 30, 46)',
-    secondary: 'rgb(91, 185, 236)',
-    onSecondary: 'rgb(4,24,37)',
-    secondaryContainer: 'rgb(227, 223, 255)',
-    onSecondaryContainer: 'rgb(19, 4, 99)',
-    tertiary: 'rgb(79, 103, 0)',
-    onTertiary: 'rgb(255, 255, 255)',
-    tertiaryContainer: 'rgb(138, 174, 30)',
-    onTertiaryContainer: 'rgb(4,24,37)',
-    error: 'rgb(138, 174, 30)',
-    onError: 'rgb(6, 35, 51)',
-    errorContainer: 'rgb(138, 174, 30)',
-    onErrorContainer: 'rgb(6, 35, 51)',
-    background: 'rgb(237, 246, 253)',
-    onBackground: 'rgb(25, 28, 30)',
-    surface: 'rgb(237, 246, 253)',
-    onSurface: 'rgb(25, 28, 30)',
-    surfaceVariant: 'rgb(221, 227, 234)',
-    onSurfaceVariant: 'rgb(65, 72, 77)',
-    outline: 'rgb(113, 120, 126)',
-    outlineVariant: 'rgb(193, 199, 206)',
-    shadow: 'rgb(0, 0, 0)',
-    scrim: 'rgb(0, 0, 0)',
-    inverseSurface: 'rgb(18, 8, 2)',
-    inverseOnSurface: 'rgb(230, 230, 225)',
-    inversePrimary: 'rgb(249, 220, 204)',
-    elevation: {
-      level0: 'transparent',
-      level1: 'rgb(237, 246, 253)',
-      level2: 'rgb(232, 240, 246)',
-      level3: 'rgb(224, 235, 243)',
-      level4: 'rgb(222, 234, 242)',
-      level5: 'rgb(217, 231, 239)',
-    },
-    surfaceDisabled: 'rgba(25, 28, 30, 0.12)',
-    onSurfaceDisabled: 'rgba(25, 28, 30, 0.38)',
-    backdrop: 'rgba(43, 49, 54, 0.4)',
-  },
-};
+import { MD3DarkTheme, MD3Theme } from 'react-native-paper';
 
 export const darkTheme: MD3Theme = {
   ...MD3DarkTheme,
